requirements.txt
accelerate==0.15.0
cupy-cuda11x==12.3.0
diffusers==0.10.2
einops==0.8.1
einops-exts==0.0.4
ema-pytorch==0.7.7
huggingface-hub==0.20.2
imageio==2.35.1
lpips==0.1.4
mediapy==1.2.4
omegaconf==2.3.0
opencv-python==4.11.0.86
PyYAML==6.0.2
scikit-image==0.21.0
scipy==1.9.3
termcolor==2.4.0
tqdm==4.67.1
wandb==0.19.11
